pyproject.toml
[project]
name = "llm7shi"
version = "0.1.0"
description = "A simplified Python library for interacting with large language models"
readme = "README.md"
license = { text = "CC0 1.0 Universal (CC0 1.0) Public Domain Dedication" }
repository = "https://github.com/7shi/llm7shi"
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.6",
    "google-genai>=1.21.1",
    "openai>=1.91.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llm7shi"]
